feat: Add background color to BigSidebar component
This commit adds a background color to the BigSidebar component in order to improve the visual appearance and provide better contrast with the content. The background color is set to "bg-gray-100/40" and the component is also rounded using the "rounded-lg" class. This enhancement enhances the overall design of the sidebar in the application.