feat: Update npm dependencies for @reactour/tour and embla-carousel-react
This commit updates the npm dependencies for @reactour/tour to version 3.7.0 and embla-carousel-react to version 8.1.6. Updating these dependencies ensures that the project is using the latest stable versions, which may include bug fixes, performance improvements, or new features. Keeping dependencies up to date is important for maintaining a secure and efficient codebase.