feat: Add NoCourseAndSubAlert component for displaying alert when user has no courses or subscriptions
The CoursesLayout component was updated to include the NoCourseAndSubAlert component. This change displays an alert message when the user does not have any courses or subscriptions, providing a clear indication of the lack of content. This improves the user experience by guiding them to take appropriate actions, such as enrolling in courses or subscribing to content.