refactor: Update package.json and package-lock.json dependencies
This commit updates the dependencies in the package.json and package-lock.json files. It adds the "framer-motion" package with version "^11.3.8" and the "mini-svg-data-uri" package with version "^1.4.4". These updates ensure that the project uses the latest stable versions of these dependencies, which may introduce new features, bug fixes, and performance improvements.