feat: Add ParticlesSection component to home page
This commit adds the ParticlesSection component to the home page. The ParticlesSection component is a React component that displays a particle effect using the Particles component from the MagicUI module. It dynamically sets the color of the particles based on the theme selected using the useTheme hook from the next-themes library. This addition enhances the visual appeal and interactivity of the home page, allowing developers to easily incorporate particle effects into their applications.